+++ pcs_test/tier0/lib/commands/sbd/test_disable_sbd.py
@@ -14,9 +14,15 @@ class DisableSbd(TestCase):
         self.corosync_conf_name = "corosync-3nodes.conf"
         self.node_list = ["rh7-1", "rh7-2", "rh7-3"]
         self.config.env.set_known_nodes(self.node_list)
+        self.cib_resources = """
+            <resources>
+                <primitive id="S-any" class="stonith" type="fence_any" />
+            </resources>
+        """
 
     def test_success(self):
         self.config.corosync_conf.load(filename=self.corosync_conf_name)
+        self.config.runner.cib.load(resources=self.cib_resources)
         self.config.http.host.check_auth(node_labels=self.node_list)
         self.config.http.pcmk.set_stonith_watchdog_timeout_to_zero(
             communication_list=[[dict(label=node)] for node in self.node_list],
@@ -49,11 +55,69 @@ class DisableSbd(TestCase):
             ]
         )
 
+    def test_no_stonith_left(self):
+        self.config.corosync_conf.load(filename=self.corosync_conf_name)
+        self.config.runner.cib.load()
+
+        self.env_assist.assert_raise_library_error(
+            lambda: disable_sbd(self.env_assist.get_env())
+        )
+        self.env_assist.assert_reports(
+            [
+                fixture.error(
+                    reports.codes.NO_STONITH_MEANS_WOULD_BE_LEFT,
+                    force_code=reports.codes.FORCE,
+                )
+            ]
+        )
+
+    def test_no_stonith_left_forced(self):
+        self.config.corosync_conf.load(filename=self.corosync_conf_name)
+        self.config.runner.cib.load()
+        self.config.http.host.check_auth(node_labels=self.node_list)
+        self.config.http.pcmk.set_stonith_watchdog_timeout_to_zero(
+            communication_list=[[dict(label=node)] for node in self.node_list],
+        )
+        self.config.http.sbd.disable_sbd(node_labels=self.node_list)
+
+        disable_sbd(
+            self.env_assist.get_env(), force_flags={reports.codes.FORCE}
+        )
+        self.env_assist.assert_reports(
+            [
+                fixture.warn(
+                    reports.codes.NO_STONITH_MEANS_WOULD_BE_LEFT,
+                ),
+                fixture.info(
+                    reports.codes.SERVICE_ACTION_STARTED,
+                    action=reports.const.SERVICE_ACTION_DISABLE,
+                    service="sbd",
+                    instance="",
+                ),
+            ]
+            + [
+                fixture.info(
+                    reports.codes.SERVICE_ACTION_SUCCEEDED,
+                    action=reports.const.SERVICE_ACTION_DISABLE,
+                    service="sbd",
+                    node=node,
+                    instance="",
+                )
+                for node in self.node_list
+            ]
+            + [
+                fixture.warn(
+                    report_codes.CLUSTER_RESTART_REQUIRED_TO_APPLY_CHANGES
+                )
+            ]
+        )
+
     def test_some_node_names_missing(self):
         self.corosync_conf_name = "corosync-some-node-names.conf"
         self.node_list = ["rh7-2"]
 
         self.config.corosync_conf.load(filename=self.corosync_conf_name)
+        self.config.runner.cib.load(resources=self.cib_resources)
         self.config.http.host.check_auth(node_labels=self.node_list)
         self.config.http.pcmk.set_stonith_watchdog_timeout_to_zero(
             communication_list=[[dict(label=node)] for node in self.node_list],
@@ -94,6 +158,7 @@ class DisableSbd(TestCase):
 
     def test_all_node_names_missing(self):
         self.config.corosync_conf.load(filename="corosync-no-node-names.conf")
+        self.config.runner.cib.load(resources=self.cib_resources)
         self.env_assist.assert_raise_library_error(
             lambda: disable_sbd(self.env_assist.get_env())
         )
@@ -112,6 +177,7 @@ class DisableSbd(TestCase):
     def test_node_offline(self):
         err_msg = "Failed connect to rh7-3:2224; No route to host"
         self.config.corosync_conf.load(filename=self.corosync_conf_name)
+        self.config.runner.cib.load(resources=self.cib_resources)
         self.config.http.host.check_auth(
             communication_list=[
                 {"label": "rh7-1"},
@@ -145,6 +211,7 @@ class DisableSbd(TestCase):
         err_msg = "Failed connect to rh7-3:2224; No route to host"
         online_nodes_list = ["rh7-2", "rh7-3"]
         self.config.corosync_conf.load(filename=self.corosync_conf_name)
+        self.config.runner.cib.load(resources=self.cib_resources)
         self.config.http.host.check_auth(
             communication_list=[
                 {
@@ -194,6 +261,7 @@ class DisableSbd(TestCase):
     def test_set_stonith_watchdog_timeout_fails_on_some_nodes(self):
         err_msg = "Error"
         self.config.corosync_conf.load(filename=self.corosync_conf_name)
+        self.config.runner.cib.load(resources=self.cib_resources)
         self.config.http.host.check_auth(node_labels=self.node_list)
         self.config.http.pcmk.set_stonith_watchdog_timeout_to_zero(
             communication_list=[
@@ -260,6 +328,7 @@ class DisableSbd(TestCase):
     def test_set_stonith_watchdog_timeout_fails_on_all_nodes(self):
         err_msg = "Error"
         self.config.corosync_conf.load(filename=self.corosync_conf_name)
+        self.config.runner.cib.load(resources=self.cib_resources)
         self.config.http.host.check_auth(node_labels=self.node_list)
         self.config.http.pcmk.set_stonith_watchdog_timeout_to_zero(
             communication_list=[
@@ -291,6 +360,7 @@ class DisableSbd(TestCase):
     def test_disable_failed(self):
         err_msg = "Error"
         self.config.corosync_conf.load(filename=self.corosync_conf_name)
+        self.config.runner.cib.load(resources=self.cib_resources)
         self.config.http.host.check_auth(node_labels=self.node_list)
         self.config.http.pcmk.set_stonith_watchdog_timeout_to_zero(
             communication_list=[[dict(label=node)] for node in self.node_list],