+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Command `pcs stonith sbd watchdog list` now prints watchdogs' identity and
   driver ([RHEL-76177])
 - Command `pcs cluster rename` for changing cluster name ([RHEL-76055])
+- Removing stonith devices or disabling SBD fails if the cluster would be left
+  with disabled SBD and no stonith devices ([RHEL-76170])
 
 ### Fixed
 - Command `pcs resource restart` allows restarting bundle instances (broken
@@ -21,6 +23,7 @@
 
 [RHEL-76055]: https://issues.redhat.com/browse/RHEL-76055
 [RHEL-76059]: https://issues.redhat.com/browse/RHEL-76059
+[RHEL-76170]: https://issues.redhat.com/browse/RHEL-76170
 [RHEL-76177]: https://issues.redhat.com/browse/RHEL-76177
 [RHEL-79055]: https://issues.redhat.com/browse/RHEL-79055
 [RHEL-79160]: https://issues.redhat.com/browse/RHEL-79160